+++ app/src/main/java/com/amazonaws/services/chime/sdkdemo/activity/MeetingActivity.kt
@@ -3,27 +3,29 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.amazonaws.services.chime.sdkdemo
+package com.amazonaws.services.chime.sdkdemo.activity
 
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.AudioVideoFacade
 import com.amazonaws.services.chime.sdk.meetings.session.CreateAttendeeResponse
 import com.amazonaws.services.chime.sdk.meetings.session.CreateMeetingResponse
 import com.amazonaws.services.chime.sdk.meetings.session.DefaultMeetingSession
-import com.amazonaws.services.chime.sdk.meetings.session.MeetingSession
 import com.amazonaws.services.chime.sdk.meetings.session.MeetingSessionConfiguration
 import com.amazonaws.services.chime.sdk.meetings.utils.logger.ConsoleLogger
 import com.amazonaws.services.chime.sdk.meetings.utils.logger.LogLevel
+import com.amazonaws.services.chime.sdkdemo.R
 import com.amazonaws.services.chime.sdkdemo.data.JoinMeetingResponse
+import com.amazonaws.services.chime.sdkdemo.fragment.DeviceManagementFragment
+import com.amazonaws.services.chime.sdkdemo.fragment.MeetingFragment
+import com.amazonaws.services.chime.sdkdemo.model.MeetingSessionModel
 import com.google.gson.Gson
 
-class InMeetingActivity : AppCompatActivity(),
+class MeetingActivity : AppCompatActivity(),
     DeviceManagementFragment.DeviceManagementEventListener,
-    RosterViewFragment.RosterViewEventListener {
+    MeetingFragment.RosterViewEventListener {
 
     private val logger = ConsoleLogger(LogLevel.DEBUG)
     private val gson = Gson()
@@ -35,13 +37,13 @@ class InMeetingActivity : AppCompatActivity(),
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_in_meeting)
-        meetingId = intent.getStringExtra(MeetingHomeActivity.MEETING_ID_KEY) as String
-        name = intent.getStringExtra(MeetingHomeActivity.NAME_KEY) as String
+        setContentView(R.layout.activity_meeting)
+        meetingId = intent.getStringExtra(HomeActivity.MEETING_ID_KEY) as String
+        name = intent.getStringExtra(HomeActivity.NAME_KEY) as String
 
         if (savedInstanceState == null) {
             val meetingResponseJson =
-                intent.getStringExtra(MeetingHomeActivity.MEETING_RESPONSE_KEY) as String
+                intent.getStringExtra(HomeActivity.MEETING_RESPONSE_KEY) as String
             val sessionConfig = createSessionConfiguration(meetingResponseJson)
             val meetingSession = sessionConfig?.let {
                 logger.info(TAG, "Creating meeting session for meeting Id: $meetingId")
@@ -72,7 +74,7 @@ class InMeetingActivity : AppCompatActivity(),
     }
 
     override fun onJoinMeetingClicked() {
-        val rosterViewFragment = RosterViewFragment.newInstance(meetingId)
+        val rosterViewFragment = MeetingFragment.newInstance(meetingId)
         supportFragmentManager
             .beginTransaction()
             .replace(R.id.root_layout, rosterViewFragment, "rosterViewFragment")
@@ -112,16 +114,5 @@ class InMeetingActivity : AppCompatActivity(),
             )
             null
         }
-    }
-
-    class MeetingSessionModel : ViewModel() {
-        private lateinit var meetingSession: MeetingSession
-
-        fun setMeetingSession(meetingSession: MeetingSession) {
-            this.meetingSession = meetingSession
-        }
-
-        val audioVideo: AudioVideoFacade
-            get() = meetingSession.audioVideo
     }
 }