+++ app/src/main/java/com/amazonaws/services/chime/sdkdemo/data/MetricData.kt
@@ -1,0 +1,11 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package com.amazonaws.services.chime.sdkdemo.data
+
+data class MetricData(
+    val metricName: String,
+    val metricValue: String
+)