+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [Unreleased]
+
+### Added
+* Added choose device list in the demo app for speaker button
+* Added metrics in the demo app
+
+### Changed
+* Changed to layout buttons instead of tabs
+* Changed the theme to support dark theme
+
 ## [0.5.3] - 2020-06-24
 
 ### Changed